Solutions after a great deal of trial and error on part two.
After following a seed through the decision tree, for each mapping, I calculate which follow up seeds would also be routed into the same category. The minimum value across all of the different maps is the number of seeds that can be skipped, as any seed within this range would be mapped along the exact same path and result in the previous location + n. This means the solution can be found through a depth first search of sorts and I can hold off buying a quantum computer for a bit longer.